refactoring utilities functions
