+++ apps/io-session-manager-oi/src/adapters/outbound/__tests__/notification-storage-queue.adapter.test.ts
@@ -22,11 +22,11 @@ const mocks = vi.hoisted(() => {
   };
 });
 
-vi.mock("../../../utils/codec.js", () => ({
+vi.mock("../../../utils/codec/index.js", () => ({
   Base64: mocks.Base64,
 }));
 
-vi.mock("../../../utils/crypto.js", () => ({
+vi.mock("../../../utils/crypto/index.js", () => ({
   Hash: mocks.Hash,
 }));
