+++ apps/io-session-manager-oi/src/adapters/outbound/notification-storage-queue.adapter.ts
@@ -2,8 +2,8 @@ import { QueueClient } from "@azure/storage-queue";
 import { FiscalCode, GenericError } from "@pagopa/hexagonal-core";
 import { err, ok, Result, ResultAsync } from "neverthrow";
 import { NotificationPort } from "../../domain/ports/outbound/notification.port.js";
-import { Base64 } from "../../utils/codec.js";
-import { Hash } from "../../utils/crypto.js";
+import { Base64 } from "../../utils/codec/index.js";
+import { Hash } from "../../utils/crypto/index.js";
 
 /**
  * NotificationStorageQueueAdapter is an implementation of the NotificationOutboundPort interface that uses Azure Storage Queue to send messages for deleting installations.
