+++ apps/io-session-manager-oi/src/utils/codec/__tests__/base64.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it } from "vitest";
-import { Base64 } from "../codec.js";
+import { encode } from "../base64.js";
 
 const decode = (value: string): string =>
   Buffer.from(value, "base64").toString("utf8");
 
-describe("Base64.encode", () => {
+describe("encode", () => {
   it.each([
     ["a string", "a string"],
     [42, "42"],
@@ -13,22 +13,22 @@ describe("Base64.encode", () => {
     [undefined, "undefined"],
     [123n, "123"],
   ])("encodes primitive %p", (value, expected) => {
-    expect(decode(Base64.encode(value))).toBe(expected);
+    expect(decode(encode(value))).toBe(expected);
   });
 
   it("encodes binary values without serializing them", () => {
     const value = Buffer.from([0, 1, 255]);
 
-    expect(Base64.encode(value)).toBe(value.toString("base64"));
+    expect(encode(value)).toBe(value.toString("base64"));
   });
 
   it("JSON serializes objects", () => {
     const value = { enabled: true, id: 42, tags: ["session"] };
 
-    expect(decode(Base64.encode(value))).toBe(JSON.stringify(value));
+    expect(decode(encode(value))).toBe(JSON.stringify(value));
   });
 
   it("serializes bigint properties as strings", () => {
-    expect(decode(Base64.encode({ id: 123n }))).toBe('{"id":"123"}');
+    expect(decode(encode({ id: 123n }))).toBe('{"id":"123"}');
   });
 });
