--- apps/io-session-manager-oi/src/utils/crypto.ts
@@ -1,7 +1,0 @@
-import { createHash } from "node:crypto";
-
-export namespace Hash {
-  export const sha256 = (value: string): string => {
-    return createHash("sha256").update(value).digest("hex");
-  };
-}
