+++ apps/io-session-manager-oi/src/utils/crypto/hash.ts
@@ -1,0 +1,5 @@
+import { createHash } from "node:crypto";
+
+export const sha256 = (value: string): string => {
+  return createHash("sha256").update(value).digest("hex");
+};
